+++ src/lib.rs
@@ -15,8 +15,8 @@ pub use quick_xml::{Reader as XmlReader, Writer as XmlWriter};
 #[cfg(feature = "download")]
 pub use taxonomy::TaxonomyLoader;
 pub use taxonomy::{
-    ArcroleType, CalculationArc, DefinitionArc, ElementDefinition, Label, LinkbaseLocator,
-    LinkbaseRef, PresentationArc, Reference, ReferencePart, RoleType, SchemaImport, SchemaInclude,
-    TaxonomySchema, TaxonomySet,
+    ArcroleType, CalculationArc, ConceptId, DefinitionArc, ElementDefinition, Label,
+    LinkbaseLocator, LinkbaseRef, PresentationArc, Reference, ReferencePart, RoleType, RoleUri,
+    SchemaImport, SchemaInclude, SchemaRefUrl, TaxonomySchema, TaxonomySet,
 };
 pub use validation::{Severity, ValidationMessage, ValidationResult};