+++ src/taxonomy/mod.rs
@@ -21,4 +21,4 @@ pub use schema::{
     ArcroleType, ElementDefinition, LinkbaseRef, RoleType, SchemaImport, SchemaInclude,
     TaxonomySchema,
 };
-pub use taxonomy_set::TaxonomySet;
+pub use taxonomy_set::{ConceptId, RoleUri, SchemaRefUrl, TaxonomySet};